+++ src/main/java/core/aop/test/DynamicInvocationHandler.java
@@ -1,5 +1,6 @@
-package core.aop;
+package core.aop.test;
 
+import core.aop.MethodMatcher;
 import net.sf.cglib.proxy.InvocationHandler;
 
 import java.lang.reflect.Method;
