+++ src/main/java/core/aop/test/HelloTarget.java
@@ -1,4 +1,4 @@
-package core.aop;
+package core.aop.test;
 
 public class HelloTarget implements Hello {
